refactor(sso): move SSO gating/bundle/identity out of openid.rs into sso_gating.rs + sso_bundle.rs (no logic change)
openid.rs keeps the shared OpenID/credential/delegation primitives used by
both direct providers and SSO. SSO-specific gating and identity resolution
move to openid/sso_gating.rs; the certified attribute bundle prepare/read/sign
paths move to openid/sso_bundle.rs. Function bodies are unchanged; only module
placement, visibility, and re-exports are adjusted so callers keep the stable
openid:: paths.
